requirements.txt
amqp==5.2.0
aniso8601==9.0.1
anyio==4.6.0
asgiref==3.8.1
async-timeout==4.0.3
billiard==4.2.1
black==24.10.0
Brotli==1.1.0
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
channels==4.1.0
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
cryptography==43.0.1
cssselect2==0.7.0
dj-rest-auth==6.0.0
Django==4.2
django-allauth==65.0.2
django-celery-beat==2.7.0
django-cors-headers==4.4.0
django-filter==24.3
django-graphql-jwt==0.4.0
django-redis==5.4.0
django-timezone-field==7.0
django_advance_thumbnail==1.0.2
django_debug_toolbar==3.8.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
exceptiongroup==1.2.2
fonttools==4.54.1
graphene==3.3
graphene-django==3.2.2
graphene-file-upload==1.3.0
graphql-core==3.2.4
graphql-relay==3.2.0
h11==0.14.0
html5lib==1.1
httptools==0.6.1
idna==3.10
inflection==0.5.1
kombu==5.4.2
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.3.6
promise==2.3
prompt_toolkit==3.0.48
psycopg2-binary==2.9.9
pycparser==2.22
pydyf==0.11.0
PyJWT==2.9.0
pyphen==0.16.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
redis==5.1.1
requests==2.32.3
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
text-unidecode==1.3
tinycss2==1.3.0
tomli==2.0.2
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.31.1
uvloop==0.20.0
vine==5.1.0
watchfiles==0.24.0
wcwidth==0.2.13
weasyprint==59.0
webencodings==0.5.1
websockets==13.1
zopfli==0.2.3
